requirements.txt
-r requirements-testing.txt
-r requirements-local.txt

pillow==4.2.1
